requirements.txt
Flask==3.0.0
edge-tts==6.1.9
Pillow==10.1.0
moviepy==1.0.3
Werkzeug==3.0.1
chardet==5.2.0
